Move new Analyzer class into core package
