+++ packages/hekla-core/src/Analyzer.js
