+++ packages/hekla-core/src/index.js
@@ -1,5 +1,5 @@
 module.exports = {
-  Analyzer: require('./analyzer'),
+  Analyzer: require('./Analyzer'),
   DependencyGraph: require('./utils/dependency-graph'),
   astUtils: require('./utils/ast-utils'),
   plugins: require('./plugins')
