+++ packages/hekla-webpack-plugin/src/HeklaWebpackPlugin.js
@@ -1,7 +1,7 @@
 const asyncLib = require('async');
 const chalk = require('chalk');
 const StickyTerminalDisplay = require('sticky-terminal-display');
-const Analyzer = require('./Analyzer');
+const Analyzer = require('hekla-core').Analyzer;
 const Module = require('./Module');
 const {
   parseAST
